requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
httpx
jinja2
openai
python-dotenv
tqdm
pdfplumber
pytesseract
Pillow
textract
# sqlite3 (built-in)
markdown
